move legacy autodiscovery to codegen
