+++ src/dbally/views/freeform/text2sql/__init__.py
@@ -1,15 +1,8 @@
-from ._autodiscovery import AutoDiscoveryBuilder, AutoDiscoveryBuilderWithLLM, configure_text2sql_auto_discovery
-from ._config import ColumnConfig, TableConfig, Text2SQLConfig, Text2SQLSimilarityType, Text2SQLTableConfig
+from ._config import ColumnConfig, TableConfig
 from ._view import BaseText2SQLView
 
 __all__ = [
-    "TableConfig",
-    "ColumnConfig",
-    "Text2SQLConfig",
-    "Text2SQLTableConfig",
-    "Text2SQLSimilarityType",
     "BaseText2SQLView",
-    "configure_text2sql_auto_discovery",
-    "AutoDiscoveryBuilder",
-    "AutoDiscoveryBuilderWithLLM",
+    "ColumnConfig",
+    "TableConfig",
 ]
