+++ src/dbally_codegen/__init__.py
@@ -1,0 +1,16 @@
+from dbally_codegen.autodiscovery import (
+    AutoDiscoveryBuilder,
+    AutoDiscoveryBuilderWithLLM,
+    configure_text2sql_auto_discovery,
+)
+
+from .config import Text2SQLConfig, Text2SQLSimilarityType, Text2SQLTableConfig
+
+__all__ = [
+    "AutoDiscoveryBuilder",
+    "AutoDiscoveryBuilderWithLLM",
+    "Text2SQLSimilarityType",
+    "Text2SQLTableConfig",
+    "Text2SQLConfig",
+    "configure_text2sql_auto_discovery",
+]
